+++ client/components/domain-search-v2/register-domain-step/index.jsx
@@ -47,23 +47,6 @@ import PropTypes from 'prop-types';
 import { stringify, parse } from 'qs';
 import { Component } from 'react';
 import { connect } from 'react-redux';
-import {
-	recordDomainAvailabilityReceive,
-	recordDomainAddAvailabilityPreCheck,
-	recordFiltersReset,
-	recordFiltersSubmit,
-	recordMapDomainButtonClick,
-	recordSearchFormSubmit,
-	recordSearchFormView,
-	recordSearchResultsReceive,
-	recordShowMoreResults,
-	recordTransferDomainButtonClick,
-	recordUseYourDomainButtonClick,
-	recordDomainClickMissing,
-	resetSearchCount,
-	enqueueSearchStatReport,
-} from 'calypso/components/domains/register-domain-step/analytics';
-import { FreeDomainForAYearPromo } from 'calypso/components/domains/wpcom-domain-search/free-domain-for-a-year-promo';
 import { getDomainsInCart, hasDomainInCart } from 'calypso/lib/cart-values/cart-items';
 import {
 	checkDomainAvailability,
@@ -84,6 +67,23 @@ import { DomainSearch } from '../__legacy/domain-search';
 import { DomainCartV2 } from '../domain-cart';
 import { DomainSearchInput } from '../domain-search-input';
 import DomainSearchResults from '../domain-search-results';
+import { FreeDomainForAYearPromo } from '../domains/wpcom-domain-search/free-domain-for-a-year-promo';
+import {
+	recordDomainAvailabilityReceive,
+	recordDomainAddAvailabilityPreCheck,
+	recordFiltersReset,
+	recordFiltersSubmit,
+	recordMapDomainButtonClick,
+	recordSearchFormSubmit,
+	recordSearchFormView,
+	recordSearchResultsReceive,
+	recordShowMoreResults,
+	recordTransferDomainButtonClick,
+	recordUseYourDomainButtonClick,
+	recordDomainClickMissing,
+	resetSearchCount,
+	enqueueSearchStatReport,
+} from './analytics';
 import {
 	getStrippedDomainBase,
 	getTldWeightOverrides,
