--- client/components/domains/domain-search-results/style.scss
@@ -1,84 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.domain-search-results__transfer-card {
-	display: flex;
-	flex-direction: column;
-	margin-bottom: 12px;
-
-	@include breakpoint-deprecated( ">480px" ) {
-		align-items: center;
-		flex-direction: row;
-	}
-}
-
-.domain-search-results__transfer-card-copy {
-	font-size: $font-body;
-
-	p {
-		margin-bottom: 0;
-	}
-
-	@include breakpoint-deprecated( ">480px" ) {
-		padding-right: 10px;
-	}
-
-	@include breakpoint-deprecated( ">660px" ) {
-		font-size: $font-title-small;
-
-		p {
-			font-size: $font-body;
-		}
-	}
-}
-
-.domain-search-results__transfer-card-link {
-	flex-shrink: 0;
-	flex-grow: 1;
-	width: 100%;
-
-	a {
-		text-align: right;
-		display: block;
-		padding: 15px 0 0;
-	}
-
-	@include breakpoint-deprecated( ">480px" ) {
-		width: auto;
-
-		a {
-			padding: 0;
-		}
-	}
-}
-
-body.is-section-signup {
-	.domain-search-results {
-		@include break-xlarge {
-			margin-left: 0;
-		}
-	}
-}
-
-.domain-search-results__domain-available-notice {
-	display: flex !important;
-	.domain-search-results__domain-available-notice-icon {
-		margin-right: 8px;
-		svg {
-			display: block;
-			margin: 0 auto;
-			fill: var(--color-neutral-20);
-			width: 20px;
-			height: 20px;
-		}
-	}
-	span {
-		font-size: $font-body-small;
-		color: var(--color-text-subtle);
-		a {
-			color: var(--color-text);
-			text-decoration: underline;
-			font-weight: 600;
-		}
-	}
-}
