--- client/components/domains/domain-skip-suggestion/style.scss
@@ -1,43 +1,0 @@
-.domain-skip-suggestion {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-
-	& .button.domain-suggestion__action.is-borderless {
-		flex: 1 0 auto;
-	}
-
-	.domain-suggestion__action-container {
-		display: flex;
-	}
-
-	.domain-suggestion__content {
-		width: auto;
-	}
-}
-
-.domain-skip-suggestion::after {
-	display: none;
-}
-
-.domain-skip-suggestion__domain-description {
-	padding-right: 1em;
-
-	@include breakpoint-deprecated( ">660px" ) {
-		width: 75%;
-	}
-
-	h3 {
-		word-break: break-all;
-	}
-
-	p {
-		color: var(--color-neutral-70);
-		font-size: $font-body-extra-small;
-		margin-bottom: 0;
-
-		@include breakpoint-deprecated( ">960px" ) {
-			margin-bottom: 8px;
-		}
-	}
-}
