--- client/components/domains/example-domain-browser/style.scss
@@ -1,16 +1,0 @@
-.example-domain-browser {
-	pointer-events: none;
-
-	svg {
-		position: relative;
-		top: 15px;
-	}
-}
-
-.example-domain-browser__protocol {
-	fill: var(--color-text-subtle);
-}
-
-.example-domain-browser__domain {
-	fill: var(--color-text);
-}
