--- client/components/domains/example-domain-suggestions/style.scss
@@ -1,59 +1,0 @@
-.example-domain-suggestions {
-	box-sizing: border-box;
-	position: relative;
-	padding-top: 16px;
-	padding-right: 348px;
-	max-width: 600px;
-	margin: 0 auto 60px;
-
-	p {
-		color: var(--color-text-inverted);
-		text-align: right;
-		margin-bottom: 8px;
-	}
-
-	a {
-		text-decoration: underline;
-		color: var(--color-text-inverted);
-	}
-
-	@include breakpoint-deprecated( "<660px" ) {
-		max-width: 100%;
-		padding: 16px;
-		margin: 0 auto;
-
-		p {
-			text-align: center;
-		}
-	}
-}
-
-.example-domain-suggestions__browser {
-	position: absolute;
-	top: 0;
-	right: 0;
-	width: 300px;
-
-	@include breakpoint-deprecated( "<660px" ) {
-		position: relative;
-		margin: 0 auto;
-	}
-}
-
-body.is-section-signup {
-	.example-domain-suggestions {
-		max-width: 678px;
-
-		p {
-			text-align: left;
-		}
-		.example-domain-suggestions__explanation {
-			color: var(--color-neutral-40);
-		}
-		.example-domain-suggestions__mapping-information {
-			a {
-				color: var(--color-neutral-60);
-			}
-		}
-	}
-}
