--- client/components/domains/featured-domain-suggestions/placeholder.scss
@@ -1,43 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.featured-domain-suggestion.featured-domain-suggestion--is-placeholder {
-	min-height: 222px;
-
-	.domain-registration-suggestion__title,
-	.domain-registration-suggestion__badges,
-	.domain-suggestion__action-container,
-	.domain-product-price,
-	.domain-registration-suggestion__title-info,
-	.domain-suggestion__action {
-		animation: loading-fade 1.6s ease-in-out infinite;
-		background-color: var(--color-neutral-0);
-		color: transparent;
-	}
-
-	.domain-registration-suggestion__title {
-		padding: 0;
-		width: 60%;
-		height: 45px;
-	}
-	.domain-registration-suggestion__title-info,
-	.domain-product-price {
-		height: 18px;
-	}
-	.domain-registration-suggestion__badges {
-		height: 22px;
-		width: 20%;
-		align-self: start !important;
-
-	}
-	.domain-suggestion__action-container {
-		margin-bottom: 6px;
-	}
-
-	.domain-suggestion__content {
-		flex-grow: 0;
-		@include break-mobile {
-			flex-grow: 1;
-		}
-	}
-}
