--- client/components/domains/featured-domain-suggestions/style.scss
@@ -1,321 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.featured-domain-suggestions {
-	display: flex;
-	flex-direction: column;
-	flex-wrap: wrap;
-	border-top: 1px solid rgba(220, 220, 222, 0.64);
-
-	@include break-small {
-		border-top: none;
-	}
-
-	&.featured-domain-suggestions--has-match-reasons .featured-domain-suggestion {
-		.domain-registration-suggestion__match-reasons {
-			min-height: 26px;
-		}
-	}
-
-	.is-section-signup & {
-		@include breakpoint-deprecated( ">480px" ) {
-			flex-direction: row;
-		}
-		display: block;
-	}
-
-	.is-section-domains & {
-		@include breakpoint-deprecated( ">800px" ) {
-			flex-direction: row;
-		}
-	}
-}
-
-.featured-domain-suggestion {
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-	margin-top: 0;
-	width: 100%;
-
-	@include break-mobile {
-		flex-direction: row;
-	}
-
-	.domain-registration-suggestion__title-info {
-		max-width: unset;
-	}
-
-	.button.domain-suggestion__action {
-		flex-grow: 0;
-		text-align: center;
-
-		.is-placeholder & {
-			animation: loading-fade 1.6s ease-in-out infinite;
-			background-color: var(--color-neutral-0);
-			border: none;
-			border-radius: 0;
-			color: transparent;
-			margin-left: 40px;
-			min-height: 26px;
-		}
-	}
-
-	.domain-suggestion__content {
-		display: flex;
-		flex-basis: auto;
-		flex-direction: column;
-		flex-grow: 1;
-		margin-top: 0;
-		flex-wrap: wrap;
-		min-height: 95px;
-		.domain-registration-suggestion__title-info {
-			width: 100%;
-		}
-		@media (min-width: 660px) {
-			flex-direction: row;
-		}
-	}
-
-	.domain-registration-suggestion__title,
-	.domain-registration-suggestion__badges,
-	.domain-product-price,
-	.domain-registration-suggestion__match-reasons {
-		flex: 0 0 auto;
-		align-self: flex-start;
-	}
-
-	.domain-registration-suggestion__title {
-		font-size: $font-title-small;
-		font-weight: 400;
-		line-height: 1.2;
-		margin-bottom: 0.25em;
-		width: auto;
-
-		@include break-mobile {
-			margin-bottom: 0;
-		}
-
-		.domain-registration-suggestion__domain-title {
-			font-size: $font-title-small;
-
-			@include break-mobile {
-				font-size: $font-title-medium;
-			}
-
-			.domain-registration-suggestion__domain-title-name {
-				color: var(--studio-gray-90);
-			}
-
-			.domain-registration-suggestion__domain-title-tld {
-				font-weight: normal;
-			}
-		}
-	}
-
-	// .card used to increase specificity
-	&.card .domain-product-price {
-		margin: 0;
-		flex-grow: 1;
-		align-items: baseline;
-		.is-section-domains & {
-			padding-left: 0;
-		}
-		&.is-free-domain {
-			padding-left: 0;
-			display: flex;
-			flex-direction: column-reverse;
-			@include break-mobile {
-				flex-direction: column;
-			}
-		}
-	}
-
-	.domain-suggestion__price-container {
-		flex-grow: 1;
-		display: flex;
-	}
-
-	.domain-registration-suggestion__badges {
-		align-self: center;
-		align-items: center;
-		display: flex;
-		font-size: $font-body-extra-small;
-		order: 2;
-
-		.progress-bar {
-			width: 25%;
-			margin-right: 1em;
-		}
-	}
-
-	.domain-registration-suggestion__match-reasons {
-		order: 3;
-	}
-
-	.domain-registration-suggestion__match-reason {
-		color: var(--color-text-subtle);
-		font-size: $font-body-small;
-		padding: 0.125em 0;
-		display: flex;
-		align-items: center;
-
-		.gridicon {
-			color: var(--color-success);
-			margin-right: 0.25em;
-		}
-	}
-
-	&.is-unavailable {
-		.domain-registration-suggestion__badges,
-		.domain-registration-suggestion__match-reasons {
-			display: none;
-		}
-	}
-}
-
-.featured-domain-suggestion.card {
-	.domain-registration-suggestion__title-wrapper {
-		flex-wrap: nowrap;
-		flex-direction: column-reverse;
-
-		@include break-xlarge {
-			flex-direction: row;
-		}
-	}
-}
-
-body.is-section-stepper .step-container .featured-domain-suggestion.card,
-body.is-section-stepper .step-container-v2 .featured-domain-suggestion.card,
-body.is-section-signup .step-wrapper .featured-domain-suggestion.card {
-	border: none;
-	border-radius: 0;
-	margin: 0;
-	align-items: center;
-	background: #fff;
-	border-bottom: 1px solid rgba(220, 220, 222, 0.64);
-	flex-direction: column-reverse;
-	@include break-mobile {
-		border-radius: 4px;
-		margin-bottom: 12px;
-		box-sizing: border-box;
-		border: 1px solid #e2e4e7;
-		background: var(--color-surface);
-	}
-
-	.domain-suggestion__action-container {
-		width: 100%;
-
-		@include break-mobile {
-			width: auto;
-		}
-	}
-
-	.button.domain-suggestion__action {
-		padding: 0.57em 1.17em;
-		border-radius: 4px;
-		font-size: 0.875rem;
-		height: auto;
-
-		@include break-mobile {
-			margin-top: 0;
-			padding: 0.65em 2.8em;
-		}
-
-		&:focus {
-			border-color: var(--color-primary);
-			box-shadow: 0 0 0 2px var(--color-primary-light);
-		}
-	}
-
-	.domain-registration-suggestion__domain-title {
-		font-size: 1.375rem; /* stylelint-disable-line scales/font-sizes */
-		line-height: 30px;
-
-		@include break-mobile {
-			font-size: 1.625rem; /* stylelint-disable-line scales/font-sizes */
-		}
-	}
-
-	.domain-registration-suggestion__title-wrapper {
-		flex-wrap: nowrap;
-		flex-direction: column-reverse;
-		margin-bottom: initial;
-
-		@include break-wide {
-			&:not(.domain-registration-suggestion__larger-domain) {
-				flex-direction: row;
-			}
-		}
-	}
-
-	.domain-registration-suggestion__title {
-		font-size: $font-title-medium;
-		padding-right: 0;
-		margin-bottom: 0;
-		flex: 0 1 auto;
-		align-self: baseline;
-	}
-
-	&.domain-suggestion.is-clickable:hover {
-		@include break-mobile {
-			border: 1px solid #646970;
-		}
-	}
-
-	&.featured-domain-suggestion--is-placeholder {
-		margin-bottom: 1px;
-		margin-top: 0;
-		border-radius: 0;
-		border-bottom: none;
-
-		@include break-mobile {
-			margin-bottom: 12px;
-			border-radius: 4px;
-			border: none;
-		}
-	}
-}
-
-body.is-section-signup .featured-domain-suggestions {
-	margin: 0;
-	border-top: 1px solid rgba(220, 220, 222, 0.64);
-
-	@include break-small {
-		margin: 0 20px;
-		border-top: none;
-	}
-
-	@include break-large {
-		margin: 0;
-	}
-}
-
-@include breakpoint-deprecated( ">660px" ) {
-	.domain-registration-suggestion__title {
-		.featured-domain-suggestions--title-in-18em & {
-			font-size: 1.8em;
-		}
-		.featured-domain-suggestions--title-in-16em & {
-			font-size: 1.6em;
-		}
-		.featured-domain-suggestions--title-in-14em & {
-			font-size: $font-title-small;
-		}
-		.featured-domain-suggestions--title-in-12em & {
-			font-size: 1.2em;
-		}
-		.featured-domain-suggestions--title-in-10em & {
-			font-size: 1em;
-		}
-	}
-
-	.featured-domain-suggestions--title-causes-overflow {
-		.featured-domain-suggestion {
-			margin-top: 0;
-		}
-		.featured-domain-suggestion .domain-suggestion__action.button.is-primary {
-			width: initial;
-		}
-	}
-}
