--- client/components/domains/register-domain-step/style.scss
@@ -1,316 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.register-domain-step:not(.register-domain-step__signup) {
-	display: block;
-	position: relative;
-	margin: 0 auto 10px;
-	padding: 16px;
-	box-sizing: border-box;
-	background: var(--color-surface);
-	box-shadow: 0 0 0 1px var(--color-border-subtle);
-
-	@include break-mobile {
-		padding: 24px;
-	}
-
-	.search-filters__dropdown-filters {
-		height: 40px;
-
-		@include break-mobile {
-			height: 48px;
-		}
-	}
-
-	.button.domain-suggestion__action {
-		padding: 0;
-		@include break-mobile {
-			min-width: 120px;
-		}
-	}
-
-	.card {
-		background: none;
-		box-shadow: none;
-		border-bottom: 1px solid var(--color-border-subtle);
-		padding-left: 0;
-		padding-right: 0;
-	}
-
-	.domain-search-results__domain-available-notice,
-	.search-component {
-		border-bottom: none;
-	}
-
-	.register-domain-step__next-page {
-		border-bottom: none;
-		padding-bottom: 0;
-		.register-domain-step__next-page-button {
-			padding-bottom: 0;
-			padding-top: 0;
-			@include break-mobile {
-				padding-top: 8px;
-				padding-bottom: 8px;
-			}
-		}
-	}
-
-	.domain-search-results__domain-available-notice {
-		padding-top: 0;
-		padding-bottom: 0;
-	}
-
-	.search-component {
-		overflow: hidden;
-		border: 1px solid var(--color-border-subtle);
-		border-radius: 4px;
-	}
-
-	.search-component.is-open {
-		border: 1px solid #a7aaad;
-		border-radius: 4px;
-		box-sizing: border-box;
-		height: 40px;
-
-		@include break-mobile {
-			height: 48px;
-		}
-	}
-
-	.search-component.is-open.has-focus {
-		border-color: #646970;
-		background: var(--color-surface);
-		box-shadow: none;
-
-		.search-component__input {
-			background: var(--color-surface);
-		}
-	}
-
-	.card.register-domain-step__search-card {
-		background-color: var(--color-surface);
-	}
-}
-
-.register-domain-step__search {
-	padding-bottom: 12px;
-
-	.card.register-domain-step__search-card {
-		padding: 0;
-		display: flex;
-		align-items: center;
-		border-bottom: none;
-	}
-
-	.search {
-		margin-bottom: 0;
-
-		&.is-refocused {
-			animation: shake 0.5s both;
-			box-shadow: 0 0 0 1px var(--color-neutral-light), 0 2px 4px var(--color-neutral-10);
-		}
-	}
-
-	&.disabled {
-		border-bottom: none; // so that bottom border is not there during google app dialog animation
-		opacity: 0.7;
-		transition: opacity, 0.3s, ease-in-out;
-	}
-
-	.search-card {
-		margin-bottom: 0;
-	}
-}
-
-.register-domain-step > .calypso-notice.register-domain-step__notice {
-	margin-bottom: 12px;
-}
-
-.register-domain-step__example-prompt {
-	font-size: 0.875rem;
-	line-height: 20px;
-	color: var(--studio-gray-50);
-	display: flex;
-	align-items: center;
-	margin: 0;
-
-	@include break-small {
-		margin: 0 20px;
-	}
-
-	@include break-large {
-		margin: 0;
-	}
-}
-
-.register-domain-step__placeholder.empty-content {
-	padding-bottom: 30px;
-}
-
-@keyframes shake {
-	0%,
-	100% {
-		transform: translate3d(0, 0, 0);
-	}
-
-	10%,
-	60% {
-		transform: translate3d(-5px, 0, 0);
-	}
-
-	30% {
-		transform: translate3d(5px, 0, 0);
-	}
-}
-
-.register-domain-step__filter-reset-notice {
-	color: var(--color-primary);
-	font-weight: 600;
-	width: 100%;
-	position: relative;
-	margin-bottom: 0;
-
-	border: 0;
-	border-radius: 0;
-
-	// from components/card/style.scss
-	box-shadow:
-		0 0 0 1px color-mix(in srgb, var(--color-neutral-10) 50%, transparent),
-		0 1px 2px var(--color-neutral-0);
-
-	// from components/domains/domain-suggestion/style
-	// NOTE: easeOutExpo easing function from http://easings.net/#easeOutExpo
-	transition: box-shadow 0.25s cubic-bezier(0.19, 1, 0.22, 1);
-
-	&:hover {
-		color: var(--color-primary);
-
-		// from components/domains/domain-suggestion/style
-		box-shadow: 0 0 0 1px var(--color-neutral-light);
-	}
-}
-
-.register-domain-step__next-page {
-	color: var(--color-link);
-	display: flex;
-	justify-content: center;
-}
-
-.register-domain-step {
-	.button.is-active {
-		background: var(--color-neutral-0);
-	}
-}
-
-button.register-domain-step__next-page-button {
-	font-size: 0.875rem;
-}
-
-body.is-section-domains {
-	.register-domain-step {
-		button.register-domain-step__next-page-button {
-			color: var(--color-link);
-			border: none;
-			box-shadow: none;
-		}
-	}
-}
-
-body.is-section-stepper,
-body.is-section-signup {
-	.domain-search-results__domain-availability .card {
-		box-shadow: none;
-		padding: 0;
-		margin-bottom: 24px;
-		margin-left: 20px;
-
-		@include break-mobile {
-			margin-left: 0;
-		}
-	}
-	.domain-search-results__domain-availability svg,
-	.register-domain-step__notice svg {
-		margin: 0;
-		padding: 0;
-	}
-	.register-domain-step__next-page {
-		background: none;
-		box-shadow: none;
-
-		button.register-domain-step__next-page-button {
-			padding: 0.57em 1.17em;
-			border-radius: 4px;
-
-			letter-spacing: 0.32px;
-			/* stylelint-disable-next-line declaration-property-unit-allowed-list */
-			line-height: 1.25rem;
-
-			@include break-mobile {
-				padding: 0.65em 2.8em;
-			}
-		}
-	}
-}
-
-.already-own-a-domain {
-	font-size: $font-body-small;
-	color: var(--color-text-subtle);
-	display: flex;
-	align-items: center;
-	margin-top: 26px;
-	svg {
-		align-self: flex-start;
-		width: 16px;
-		height: 16px;
-		fill: var(--color-text-subtle);
-		margin: 3px 8px 0;
-		@include break-mobile {
-			margin-left: 0;
-		}
-	}
-	.button-plain {
-		color: var(--color-link);
-	}
-}
-.register-domain-step__placeholder {
-	display: none;
-
-	@include break-large {
-		display: block;
-	}
-}
-
-.is-domain-plan-package-flow {
-
-	.register-domain-step {
-		box-shadow: none;
-		background: var(--color-body-background);
-	}
-
-	.search-filters__dropdown-filters {
-		border-color: transparent;
-	}
-
-	.register-domain-step__domains-quickfilter-group {
-
-		@include break-large {
-			display: flex;
-		}
-
-		.components-button {
-			&.is-pressed {
-				&::before {
-					border-radius: 4px;
-				}
-			}
-		}
-	}
-}
-
-.search-component .search-filters__dropdown-filters button {
-	background-color: transparent;
-	border-color: transparent;
-	border-radius: 0;
-	border-right: none;
-	border-left: none;
-}
