--- client/components/domains/search-filters/index.js
@@ -1,6 +1,0 @@
-import DropdownFilters from './dropdown-filters';
-import FilterResetNotice from './filter-reset-notice';
-
-import './style.scss';
-
-export { DropdownFilters, FilterResetNotice };
