--- client/components/domains/search-filters/style.scss
@@ -1,209 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.search-filters__dropdown-filters {
-	align-items: center;
-	height: 40px;
-	z-index: z-index("root", ".search");
-	margin-left: 12px;
-
-	@include break-mobile {
-		margin-left: 20px;
-		height: 48px;
-	}
-
-	.button {
-		align-items: center;
-		display: flex;
-		font-weight: normal;
-		height: 100%;
-		justify-content: center;
-		padding: 0 1em;
-		transition: 0.1s all linear;
-		width: 100%;
-		color: var(--color-text);
-
-		@include break-mobile {
-			min-width: 96px;
-		}
-
-		background-color: var(--studio-white);
-
-		border: 1px solid var(--color-neutral-20);
-		border-radius: 4px;
-
-		&.is-borderless {
-			&:hover,
-			&:focus {
-				border-color: var(--color-neutral);
-			}
-		}
-
-		.material-icon {
-			top: 0;
-			margin-right: 4px;
-
-			@include breakpoint-deprecated( "<660px" ) {
-				top: 2px;
-			}
-		}
-
-		.count {
-			margin-left: 6px;
-		}
-
-		.search-filters__dropdown-filters-button-text {
-			white-space: nowrap;
-		}
-	}
-}
-
-.search-filters__popover {
-	width: 28em;
-
-	// Use increased specificty to override default z-index for popovers
-	&.popover {
-		z-index: z-index(".popover", ".search-filters__popover");
-	}
-
-	.popover__inner {
-		padding: 2em;
-	}
-
-	.form-fieldset {
-		text-align: left;
-
-		&:last-child {
-			margin-bottom: 0;
-		}
-	}
-
-	.validation-fieldset {
-		margin-bottom: 1em;
-	}
-	.validation-fieldset__validation-message {
-		min-height: initial;
-	}
-}
-
-.search-filters__buttons {
-	display: flex;
-	flex-flow: row;
-	overflow: hidden;
-	justify-content: space-between;
-
-	.button {
-		flex: 1 0 auto;
-		margin-left: 1em;
-		&:first-child {
-			margin-left: 0;
-		}
-
-		&.search-filters__button--is-placeholder {
-			animation: loading-fade 1.6s ease-in-out infinite;
-			background-color: var(--color-neutral-0);
-			color: transparent;
-			height: 40px;
-		}
-	}
-
-	// Increase specificity to override button styles in signup
-	body.is-section-signup .layout & {
-		button.search-filters__popover-button {
-			font-size: $font-body-small;
-			padding-top: 0.5em;
-			padding-bottom: 0.5em;
-		}
-	}
-}
-
-.search-filters__popover-button .gridicon {
-	margin-left: 2px;
-}
-
-.search-filters__tld-filters {
-	.token-field__token {
-		margin: 4px 8px 4px 0;
-	}
-	.token-field__token-text,
-	.token-field__remove-token {
-		background: var(--color-surface);
-		color: var(--color-neutral-50);
-		border: 1px solid var(--color-neutral-10);
-	}
-	.token-field__token-text {
-		border-right: none;
-	}
-	.token-field__remove-token {
-		border-left: none;
-	}
-
-	.token-field__input-container {
-		margin: 5px 5px 5px 16px;
-	}
-
-	input[type="text"].token-field__input {
-		margin-left: 0;
-		padding-left: 0;
-	}
-
-	.token-field__suggestions-list.is-expanded {
-		padding-top: 0;
-		overflow-y: scroll;
-		max-height: 18.5em;
-	}
-
-	.token-field__suggestion-match {
-		color: inherit;
-	}
-
-	.token-field__suggestion {
-		padding: 6px 8px 6px 16px;
-		color: var(--color-neutral-70);
-		&.is-selected span {
-			color: var(--color-text-inverted);
-		}
-		&:first-child {
-			padding-top: 12px;
-		}
-	}
-}
-
-body.is-section-signup {
-	.search-filters-extensions__popover {
-		@include break-medium {
-			transform: translateX(65px);
-		}
-	}
-
-	.search-filters__popover-button {
-		order: 7;
-		margin-left: 1em;
-		background: #fdfdfd;
-		border: 1px solid #d5d5d7;
-		border-radius: 4px;
-		font-weight: 500;
-		color: var(--studio-gray-60);
-
-		&.is-active {
-			background: initial;
-			font-weight: 500;
-			border-color: var(--studio-gray-50);
-			color: #2b2d2f;
-		}
-	}
-	.search-filters__buttons {
-		margin: 0 20px 0;
-		padding: 0;
-		box-shadow: none;
-
-		@include break-large {
-			margin-left: 0;
-			margin-right: 0;
-		}
-	}
-
-	.search-filters__buttons .button {
-		border-radius: 4px;
-	}
-}
