--- client/components/domains/side-explainer/style.scss
@@ -1,71 +1,0 @@
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
-
-.side-explainer {
-	width: 100%;
-
-	.side-explainer__content {
-		display: flex;
-		flex-direction: column;
-		gap: 40px;
-	}
-
-	.side-explainer__separator {
-		height: 1px;
-		background: var( --studio-gray-5 );
-		width: 100%;
-	}
-
-	.side-explainer__subtitle-free-domain {
-		color: var(--studio-gray-60);
-		/* stylelint-disable-next-line declaration-property-unit-allowed-list */
-		line-height: 1.25rem;
-		letter-spacing: -0.16px;
-		font-size: 0.875rem;
-	}
-
-	.side-explainer__title {
-		font-size: 1.25rem;
-		letter-spacing: 0.2px;
-		/* stylelint-disable-next-line declaration-property-unit-allowed-list */
-		line-height: 1.625rem;
-		color: var(--studio-gray-90);
-		text-wrap: balance;
-	}
-
-	.side-explainer__subtitle {
-		color: var(--studio-gray-60);
-		/* stylelint-disable-next-line declaration-property-unit-allowed-list */
-		line-height: 1.25rem;
-		letter-spacing: -0.16px;
-		padding-top: 12px;
-		font-size: 0.875rem;
-	}
-
-	.side-explainer__subtitle-2 {
-		margin-top: 1.5em;
-	}
-
-	.side-explainer__cta {
-		font-size: 0.875rem;
-		line-height: 20px;
-		padding-top: 12px;
-		color: #1d2327;
-		letter-spacing: -0.16px;
-
-		.side-explainer__cta-text {
-			text-decoration: underline;
-			cursor: pointer;
-			font-weight: 500;
-			color: var(--studio-black);
-
-			&:focus {
-				border-color: var(--color-primary);
-				box-shadow: 0 0 0 2px var(--color-primary-light);
-			}
-		}
-	}
-	.is-loading {
-		@include placeholder( --color-neutral-10 );
-	}
-}
